cujac_half now compiles, rt exception where defect is nan/infinite

+++ opm/simulators/linalg/PreconditionerFactory_impl.hpp
@@ -46,6 +46,8 @@
 
 #include <config.h>
 
+#include <cuda_fp16.h>
+
 // Include all cuistl/GPU preconditioners inside of this headerfile
 #include <opm/simulators/linalg/PreconditionerFactoryGPUIncludeWrapper.hpp>
 
@@ -642,9 +644,9 @@ struct StandardPreconditioners<Operator, Dune::Amg::SequentialInformation> {
         F::addCreator("CUJacHalf", [](const O& op, const P& prm, const std::function<V()>&, std::size_t) {
             const double w = prm.get<double>("relaxation", 1.0);
             using block_type = typename V::block_type;
-            using VTo = Dune::BlockVector<Dune::FieldVector<float, block_type::dimension>>;
-            using matrix_type_to = typename Dune::BCRSMatrix<Dune::FieldMatrix<float, block_type::dimension, block_type::dimension>>;
-            using CuJac_t = typename cuistl::CuJac<matrix_type_to, cuistl::CuVector<float>, cuistl::CuVector<float>>;
+            using VTo = Dune::BlockVector<Dune::FieldVector<__half, block_type::dimension>>;
+            using matrix_type_to = typename Dune::BCRSMatrix<Dune::FieldMatrix<__half, block_type::dimension, block_type::dimension>>;
+            using CuJac_t = typename cuistl::CuJac<matrix_type_to, cuistl::CuVector<__half>, cuistl::CuVector<__half>>;
             using Adapter = typename cuistl::PreconditionerAdapter<VTo, VTo, CuJac_t>;
             using Converter = typename cuistl::PreconditionerConvertFieldTypeAdapter<Adapter, M, V, V>;
             auto converted = std::make_shared<Converter>(op.getmat());
